added dictionary module and refactor code
